Доп ветка. Финал.

+++ src/main/java/com/csa21472001/CalculatorService.java
@@ -1,12 +1,12 @@
 package com.csa21472001;
 public interface CalculatorService {
     String hello();
-     int sum(int num1, int num2);
+     long sum(int num1, int num2);
 
-     int minus(int num1, int num2);
+     long minus(int num1, int num2);
 
-     int multiply(int num1, int num2);
+     long multiply(int num1, int num2);
 
-     float divide(int num1, int num2);
+     double divide(int num1, int num2);
 
 }
