+++ src/main/java/com/csa21472001/CalculatorServiceImpl.java
@@ -7,23 +7,23 @@ public class CalculatorServiceImpl implements CalculatorService {
         return "<b>Добро пожаловать в калькулятор</b>";
     }
     @Override
-    public int sum(int num1, int num2) {
-        return num1 + num2;
+    public long sum(int num1, int num2) {
+        return (long)num1 + num2;
     }
     @Override
-    public int minus(int num1, int num2) {
-        return num1 - num2;
+    public long minus(int num1, int num2) {
+        return (long)num1 - num2;
     }
     @Override
-    public int multiply(int num1, int num2) {
-        return num1 * num2;
+    public long multiply(int num1, int num2) {
+        return (long)num1 * num2;
     }
     @Override
-    public float divide(int num1, int num2) {
+    public double divide(int num1, int num2) {
         if (num2 == 0) {
             throw new IllegalArgumentException();
         }
-        return (float) num1 / num2;
+        return (double) num1 / num2;
     }
 
 }
